third_party/WebKit/LayoutTests/http/tests/serviceworker/chromium/register-link-header.html

<!DOCTYPE html>
<!-- FIXME: Move this test out of chromium/ when PHP is no longer needed
     to set the Link header (crbug.com/347864).
-->
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/test-helpers.js"></script>
<body>
<script>
promise_test(function(t) {
    var scope = normalizeURL('resources/blank.html?fetch');
    var header = '<empty-worker.js>; rel=serviceworker; scope="' + scope + '"';
    var resource = 'resources/link-header.php?Link=' +
        encodeURIComponent(header);
    return with_iframe(scope)
      .then(frame =>
          Promise.all([frame.contentWindow.navigator.serviceWorker.ready,
                      fetch(resource)]))
      .then(([registration, response]) => {
          assert_equals(registration.scope, scope);
        });
  }, 'fetch can trigger service worker installation');

promise_test(function(t) {
    var scope = normalizeURL('resources/blank.html?iframe');
    var header = '<empty-worker.js>; rel=serviceworker; scope="' + scope + '"';
    var resource = 'resources/link-header.php?Link=' +
        encodeURIComponent(header);
    return with_iframe(scope)
      .then(frame =>
          Promise.all([frame.contentWindow.navigator.serviceWorker.ready,
                       with_iframe(resource)]))
      .then(([registration, frame]) => {
          assert_equals(registration.scope, scope);
        });
  }, 'An iframe can trigger service worker installation');

promise_test(function(t) {
    var scope = normalizeURL('resources/blank.html?css');
    var header = '<empty-worker.js>; rel=serviceworker; scope="' + scope + '"';
    var resource = 'resources/link-header.php?Link=' +
        encodeURIComponent(header);
    return with_iframe(scope)
      .then(frame => {
          var link = document.createElement('link');
          link.setAttribute('rel', 'stylesheet');
          link.setAttribute('type', 'text/css');
          link.setAttribute('href', resource);
          document.getElementsByTagName('head')[0].appendChild(link);
          return frame.contentWindow.navigator.serviceWorker.ready;
        })
      .then(registration => {
          assert_equals(registration.scope, scope);
        });
  }, 'A stylesheet can trigger service worker installation');

</script>
